extract FormViewHolder from Form object
